requirements.txt
absl-py==2.4.0
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
jax==0.4.38
jaxlib==0.4.38
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.21
ml_dtypes==0.5.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
protobuf==4.25.8
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-rtmidi==1.5.8
scipy==1.17.0
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.5
